live share

+++ cryptonation/app/Http/Controllers/CProduct.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Product;
-use Facade\FlareClient\Stacktrace\File;
+use Exception;
 use Illuminate\Http\Request;
 
 class CProduct extends Controller
@@ -16,17 +16,13 @@ class CProduct extends Controller
         ]);
         $imagename=time(). '-'.$req->name . '.' .$req->image->extension();
         $req->image->move(public_path('products'),$imagename);
-        if($req->status=='on'){
-            $status=1;
-        }else{
-            $status=0;
-        }
    $result= Product::insert([
         'prod_name'=>$req->name,
         'prod_price'=>$req->price,
         'prod_description'=>$req->description,
         'prod_img_url'=>$imagename,
-        'prod_status'=>$status,
+        'prod_status'=>$req->status,
+        'trending'=>$req->trending,
         'fk_cat_id'=>$req->category
     ]);
         if($result==1){
@@ -65,30 +61,31 @@ function edit_product($id){
 }
 
 function update_product(Request $req){
+        $oldimage=Product::where('prod_id',$req->id)->first('prod_img_url');
+    if($req->hasFile('image')){
         $req->validate([
             'image'=>'mimes:jpg,png,jpeg|max:5048'
         ]);
         $imagename=time(). '-'.$req->name . '.' .$req->image->extension();
-        $req->image->move(public_path('products'),$imagename);
-            if($req->status=='on'){
-            $status=1;
-        }else{
-            $status=0;
+        $res=$req->image->move(public_path('products'),$imagename);
+        try{
+        unlink(public_path('products/'.$oldimage->prod_img_url));
         }
-        $oldimage=Product::where('prod_id',$req->id)->first('prod_img_url');
+        catch(Exception $ex){
+
+        }
+    }
         $result=Product::where('prod_id', $req->id)
             ->update([
                     'prod_name'=>$req->name,
                     'prod_price'=>$req->price,
                     'prod_description'=>$req->description,
-                    'prod_img_url'=>$imagename ? : $oldimage,
-                    'prod_status'=>$status,
+                    'prod_img_url'=>(!empty($imagename) ? $imagename : $oldimage),
+                    'prod_status'=>$req->status,
+                    'trending'=>$req->trending,
                     'fk_cat_id'=>$req->category
             ]);
-                if($result==1){
-                    if($oldimage->prod_img_url){
-    unlink(public_path('products/'.$oldimage->prod_img_url));
-                    }
+                if($result){
     session(['res' => 'success']);
     session(['result' => "Successfully added"]);
     }else{
